refactor: Move functions in map.c into check.c file
