+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: sokim <sokim@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/28 11:41:26 by sokim             #+#    #+#             */
-/*   Updated: 2022/08/30 11:50:47 by sokim            ###   ########.fr       */
+/*   Updated: 2022/08/30 13:21:06 by sokim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,6 +65,7 @@ void	init_info(t_info *info);
  * Check functions
  */
 void	check_map_validation(t_info *info);
+int		check_map_contents(t_info *info, char *line);
 
 /*
  * Exit functions
